fix: move to correct directory
