+++ common-module/build.gradle
@@ -25,9 +25,8 @@ dependencies {
     implementation 'javax.validation:validation-api:2.0.1.Final'
 
     runtimeOnly 'mysql:mysql-connector-java'
+    compile 'com.h2database:h2'
 
-    compileOnly 'org.projectlombok:lombok'
-    annotationProcessor 'org.projectlombok:lombok'
     testImplementation('org.springframework.boot:spring-boot-starter-test') {
         exclude group: 'org.junit.vintage', module: 'junit-vintage-engine'
     }
