+++ common-module/src/main/java/com/yapp/crew/domain/model/Message.java
@@ -1,6 +1,6 @@
-package domain.model;
+package com.yapp.crew.domain.model;
 
-import domain.status.ReportType;
+import com.yapp.crew.domain.status.MessageType;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -13,21 +13,29 @@ import javax.persistence.Id;
 import javax.persistence.ManyToOne;
 
 @Entity
-public class Report extends BaseEntity {
+public class Message extends BaseEntity {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   private Long id;
 
-  @ManyToOne(fetch = FetchType.LAZY)
-  private User reporter;
-
-  @ManyToOne(fetch = FetchType.LAZY)
-  private User reported;
-
-  @Enumerated(value = EnumType.ORDINAL)
-  private ReportType type;
-
   @Column(nullable = false)
   private String content;
+
+  @Column(nullable = false)
+  @Enumerated(value = EnumType.ORDINAL)
+  private MessageType type;
+
+  @Column(name = "is_read", nullable = false)
+  private boolean isRead = false;
+
+  @ManyToOne(fetch = FetchType.LAZY)
+  private User sender;
+
+  @ManyToOne(fetch = FetchType.LAZY)
+  private ChatRoom chatRoom;
+
+  public Message() {
+
+  }
 }
