+++ common-module/src/main/java/com/yapp/crew/domain/status/ChatRoomStatus.java
@@ -1,4 +1,4 @@
-package domain.status;
+package com.yapp.crew.domain.status;
 
 public enum ChatRoomStatus {
   ACTIVE(0),
