+++ common-module/src/main/java/com/yapp/crew/domain/status/CityType.java
@@ -1,4 +1,4 @@
-package domain.status;
+package com.yapp.crew.domain.status;
 
 public enum CityType {
   JONGNO_GU("종로구"),
