+++ common-module/src/main/java/com/yapp/crew/domain/status/MessageType.java
@@ -1,4 +1,4 @@
-package domain.status;
+package com.yapp.crew.domain.status;
 
 public enum MessageType {
   ENTER(0, "채팅방 입장"),
