+++ common-module/src/main/java/com/yapp/crew/domain/status/ReportType.java
@@ -1,4 +1,4 @@
-package domain.status;
+package com.yapp.crew.domain.status;
 
 public enum ReportType {
   NON_MANNERS(1, "욕설 및 비매너 행위"),
