+++ common-module/src/main/resources/application.yml
@@ -1,17 +1,33 @@
 spring:
+  profiles:
+    active: default
+
+  devtools:
+    restart:
+      enabled: true
+
   h2:
     console:
       enabled: true
+
   datasource:
-    url: jdbc:mysql://localhost:3306/test?useSSL=false&characterEncoding=UTF-8&serverTimezone=UTC&createDatabaseIfNotExist=true
-    username: root
-    password:
-    driver-class-name: com.mysql.cj.jdbc.Driver
+    url: jdbc:h2:mem:testdb
+
+  data:
+    jpa:
+      repositories:
+        bootstrap-mode: default
+
   jpa:
-    database-platform: org.hibernate.dialect.MySQL5Dialect
     show-sql: true
-    hibernate:
-      ddl-auto: create
     properties:
       hibernate:
+        generate_statistics: true
         format_sql: true
+
+logging:
+  level:
+    org:
+      hibernate:
+        stat: debug
+        type: trace
